Update simulation.md
`genion` overwrites the topology correctly if option "-p" is specified by modifying the [molecules] section at the end of the topology file. I included this option in the command line and removed all the tasks related to the manual editing step. The new topology also contains the charges of the ions and not only "NA" and "CL" so I also took out the pic of the file